+++ package.json
@@ -19,9 +19,11 @@
     "postinstall": "playwright install"
   },
   "dependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "classnames": "2.5.1",
     "react": "19.1.0",
-    "react-dom": "19.1.0"
+    "react-dom": "19.1.0",
+    "react-router-dom": "^7.8.1"
   },
   "devDependencies": {
     "@eslint/js": "9.25.0",
